TESTS: Add allow-axfr tests
