+++ end-to-end-tests/requirements.txt
@@ -1,1 +1,2 @@
 nose>=1.3.7
+requests
